// Boardlight/src/api/generated/models/authIndexPermissionsResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: 4.1.0
 */

export interface AuthIndexPermissionsResponse {
  dangerous?: boolean;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  name?: string | null;
}