// Boardlight/src/api/generated/models/authIndexUserGroupsResponse.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Blueboard
 * OpenAPI spec version: 4.1.0
 */

export interface AuthIndexUserGroupsResponse {
  createdAt?: string;
  id?: number;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  permissions?: string[] | null;
  updatedAt?: string;
}